requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0

pydantic==1.10.13
pyyaml==6.0.1

numpy==1.23.5
pandas==2.2.1
scikit-learn==1.1.3

tensorflow==2.13.1
